data-pipeline/cloud_functions/data_cleaning_pipeline/requirements.txt
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.4
python-dateutil==2.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
